Updated code highlighting for org-mode
Earlier code blocks in this journal were still using back-tick characters, `,
rather than the org-mode tilde, `, as I started writing this journal in
markdown.

Selecting one of the ` characters and using iedit to select all matches made it
easy to replace all the back-ticks with tilde with a single R ~ combo.