Day 65: Clojure coaching and Practicalli Clojure book updates
Coaching this evening, revisiting the Hoplon project but unfortunate it doesnt
seem to run from CIDER. Not sure why its broken.

Discussed the libraries that have been moved into the clj-commons Github
organisation and submitted an issue to add the Github pages link to the
description of the clj-commons.github.io repository.

Discussed Finite state machines in Clojure and libraries that help define finite
state machines.  Finite state machines libraries include:
* http://jedahu.github.io/spaghetti/
* https://github.com/ztellman/automat

A new idea for a future coaching session is to try the duct project for building
a server side application and also use a Finite State Machine to model the UI.

More work on the Practicalli Clojure book.