Day 68 - job interviews
Today was the first job interview I have had for 15 months and it was a great
interview, I really enjoyed it. There are still a few more steps in the process,
but a very good way to start and helps build my confidence.

More work on the code examples in the Clojure->Code repository to support the
Practicalli Clojure book and a first draft of the chapter on Macros.  There is
much more I can add to the discussion of macros and how to write them, but for
now its important to define when to use them and the constraints around their
use.

* ClojureBridge London exercises and solution discussions
* Add a section on Macros to Practicalli Clojure