Organising ClojureBridge @ Functional Works
I started creating a Meetup page, but we always have the problem of how many
separate events to we create.  So, this time around we are going to try using
BridgTroll to manage the sign up process.

I contacted ClojureBridge global team and will get the training needed to manage
our events on BridgeTroll.